+++ packages/web/components/TextEditor.tsx
@@ -45,6 +45,7 @@ interface Props {
   onEvaluateCode?: (args: EvaluateCodeArgs) => void;
   onEvaluateRemoteCode?: (args: EvaluateRemoteCodeArgs) => void;
   onCursorActivity?: (args: CursorActivityArgs) => void;
+  onToggleEditorVisible?: () => void;
 }
 
 const modesByTarget = {
@@ -112,9 +113,10 @@ const Description = ({ editorId, target }) => (
 class TextEditor extends Component<Props, {}> {
   static defaultProps = {
     target: "default",
-    onEvaluateCode: () => { },
-    onEvaluateRemoteCode: () => { },
-    onCursorActivity: () => { },
+    onEvaluateCode: () => {},
+    onEvaluateRemoteCode: () => {},
+    onCursorActivity: () => {},
+    onToggleEditorVisible: () => {},
   };
 
   cm: any;
@@ -150,7 +152,7 @@ class TextEditor extends Component<Props, {}> {
     const { editor } = this.cm;
     const currentLine = editor.getCursor().line;
     const content = `${editor.getValue()}\n`;
-    const lines = content.split("\n")
+    const lines = content.split("\n");
 
     let code = "";
     let start = false;
@@ -196,8 +198,8 @@ class TextEditor extends Component<Props, {}> {
     const curBlock =
       curBlocks.length > 0 &&
       curBlocks.reduce((prev, curr) => {
-      return prev[0] < curr[0] ? prev : curr;
-    });
+        return prev[0] < curr[0] ? prev : curr;
+      });
     // console.log("Current block:", curBlock);
 
     if (curBlock) {
@@ -257,7 +259,12 @@ class TextEditor extends Component<Props, {}> {
     this.evaluate(content, 0, content.length, locally);
   };
 
-  evaluate(body: string, fromLine: number = -1, toLine: number = -1, locally: boolean = false) {
+  evaluate(
+    body: string,
+    fromLine: number = -1,
+    toLine: number = -1,
+    locally: boolean = false
+  ) {
     const { editorId, target, onEvaluateCode, sessionClient } = this.props;
 
     // console.debug("texteitor.evalute: locally = ", locally)
@@ -295,6 +302,11 @@ class TextEditor extends Component<Props, {}> {
     this.evaluate(this.freeAllSoundCode());
   };
 
+  toggleEditorVisible = () => {
+    const { onToggleEditorVisible } = this.props;
+    onToggleEditorVisible && onToggleEditorVisible();
+  };
+
   render() {
     const { editorId, isHalfHeight, target, readonly } = this.props;
 
@@ -310,7 +322,7 @@ class TextEditor extends Component<Props, {}> {
     // Replace shortkeys when using Mercury
     // Because Mercury always replaces the entire code with the newly
     // executed page. No per-line evaluation
-    if (target === 'mercury') {
+    if (target === "mercury") {
       defaultExtraKeys = {
         ...{
           "Shift-Enter": () => this.evaluateAll(false),
@@ -320,7 +332,7 @@ class TextEditor extends Component<Props, {}> {
           "Shift-Alt-Enter": () => this.evaluateAll(true),
           "Ctrl-Alt-Enter": () => this.evaluateAll(true),
           "Cmd-Alt-Enter": () => this.evaluateAll(true),
-        }
+        },
       };
     }
 
@@ -328,6 +340,7 @@ class TextEditor extends Component<Props, {}> {
       "Ctrl-.": this.freeAllSound,
       "Cmd-.": this.freeAllSound,
       "Alt-.": this.freeAllSound,
+      "Shift-Ctrl-H": this.toggleEditorVisible,
       ...defaultExtraKeys,
     };
 
@@ -342,7 +355,7 @@ class TextEditor extends Component<Props, {}> {
     };
 
     if (readonly) {
-      options['readOnly'] = "nocursor"
+      options["readOnly"] = "nocursor";
     }
 
     // const { className: heightClassName, styles: heightStyles } = css.resolve`