+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* web: Add key binding Ctrl+Shift+H to hide or show editors
+
 ### Changed
 
 * web: Replace yjsdemo WebSocket Yjs server for an embedded WS server